pyproject.toml
[tool.poetry]
name = "cloudtools"
version = "0.1.0"
description = ""
authors = ["Thinmy Patrick Alves <thinmy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.24.61"
prettytable = "^3.4.0"
click = "^8.1.3"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
